changes for refactoring
