--- mancala-api/src/main/java/org/example/domain/GameState.java
@@ -1,16 +1,0 @@
-package org.example.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NonNull;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class GameState {
-    @NonNull
-    private final String gameId;
-    @NonNull
-    private final List<playerstate1> playerstate1;
-}
