--- mancala-api/src/main/java/org/example/domain/PlayerTurn.java
@@ -1,6 +1,0 @@
-package org.example.domain;
-
-public enum PlayerTurn {
-    PLAYER_A,
-    PLAYER_B;
-}
