--- mancala-api/src/main/java/org/example/domain/playerstate1.java
@@ -1,23 +1,0 @@
-package org.example.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NonNull;
-import org.example.domain.game.core.model.Pit;
-
-import java.util.List;
-import java.util.UUID;
-
-@AllArgsConstructor
-@Getter
-public class playerstate1 {
-
-    @NonNull
-    private final UUID playerID;
-    @NonNull
-    private final boolean playerTurn;
-    @NonNull
-    private final List<Pit> pitState;
-
-
-}
